requirements.txt
aniso8601==8.0.0
apt-xapian-index==0.47
apturl==0.5.2
asn1crypto==0.24.0
Babel==2.8.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
future==0.18.2
httplib2==0.9.2
idna==2.6
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mate-hud==18.4.8
mate-tweak==18.4.16
memory-profiler==0.55.0
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.6.7
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyoa==0.115
pyparsing==2.4.6
PyQt5==5.12
PyQt5-sip==4.19.19
PyQtChart==5.12.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-xapp==1.0.0
python-xlib==0.20
pytz==2020.1
pyxdg==0.25
PyYAML==5.2
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
serial==0.0.97
sessioninstaller==0.0.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.13.0
SQLAlchemy==1.3.17
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==1.0.1
xkit==0.0.0
zope.interface==4.3.2
